# ui/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Refactored free-fs with a subject-match-java style layered architecture.",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://192.168.164.18:30067"
    }
  ],
...
  "paths": {
...
    "/plans/pages": {
      "get": {
        "summary": "分页获取套餐列表",
        "deprecated": false,
        "tags": [
          "SubscriptionPlanController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultSubscriptionPlanDTO"
                }
              }
            }
          }
        },
        "operationId": "listSubscriptionPlanPages",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "number"
            },
            "in": "query",
            "description": "No comments found.\nValidation[Min(value=1, message=pageNo必须大于0)]",
            "required": false,
            "example": 0,
            "name": "pageNo"
          },
          {
            "schema": {
              "format": "int64",
              "type": "number"
            },
            "in": "query",
            "description": "No comments found.\nValidation[Min(value=1, message=    pageSize必须大于0)]",
            "required": false,
            "example": 0,
            "name": "pageSize"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "No comments found.",
            "required": false,
            "name": "planName"
          }
        ]
      }
    },
    "/plans/info/{id}": {
      "get": {
        "summary": "获取套餐详细信息",
        "deprecated": false,
        "tags": [
          "SubscriptionPlanController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultSubscriptionPlanDTO"
                }
              }
            }
          }
        },
        "operationId": "getSubscriptionPlanById",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "number"
            },
            "in": "path",
            "description": "套餐ID",
            "required": true,
            "example": 0,
            "name": "id"
          }
        ]
      }
    },
    "/plans": {
      "post": {
        "summary": "添加套餐",
        "deprecated": false,
        "tags": [
          "SubscriptionPlanController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscriptionPlanAddCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultVoid"
                }
              }
            }
          }
        },
        "operationId": "addSubscriptionPlan",
        "parameters": []
      },
      "put": {
        "summary": "编辑套餐",
        "deprecated": false,
        "tags": [
          "SubscriptionPlanController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscriptionPlanEditCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultVoid"
                }
              }
            }
          }
        },
        "operationId": "editSubscriptionPlan",
        "parameters": []
      }
    },
    "/plans/{id}": {
      "delete": {
        "summary": "根据ID删除套餐",
        "deprecated": false,
        "tags": [
          "SubscriptionPlanController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultVoid"
                }
              }
            }
          }
        },
        "operationId": "deleteSubscriptionPlanById",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "number"
            },
            "in": "path",
            "description": "套餐ID",
            "required": true,
            "example": 0,
            "name": "id"
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SubscriptionPlanEditCommand": {
        "type": "object",
        "properties": {
          "id": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": 0
          },
          "planCode": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "planName": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "description": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "storageQuotaGb": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer",
            "example": 0
          },
          "maxFiles": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer",
            "example": 0
          },
          "maxFileSize": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": 0
          },
          "bandwidthQuota": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": 0
          },
          "price": {
            "format": "double",
            "description": "No comments found.",
            "type": "number",
            "example": 0.0
          },
          "isActive": {
            "format": "int32",
            "description": "No comments found.\nValidation[Min(value=0, message=isActive只能是0或1); Max(value=1, message=isActive只能是0或1)]",
            "type": "integer",
            "example": 0
          },
          "isDefault": {
            "format": "int32",
            "description": "No comments found.\nValidation[Min(value=0, message=isDefault只能是0或1); Max(value=1, message=isDefault只能是0或1)]",
            "type": "integer",
            "example": 0
          },
          "sortOrder": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer",
            "example": 0
          }
        },
        "required": [
          "id",
          "planCode",
          "planName",
          "storageQuotaGb",
          "maxFiles",
          "maxFileSize",
          "bandwidthQuota",
          "price",
          "isActive",
          "isDefault",
          "sortOrder"
        ]
      },
...
      "PageResultSubscriptionPlanDTO": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "totalCount": {
            "format": "int64",
            "description": "总数",
            "type": "number",
            "example": 0
          },
          "pageSize": {
            "format": "int32",
            "description": "页数",
            "type": "integer",
            "example": 0
          },
          "pageIndex": {
            "format": "int32",
            "description": "当前页数",
            "type": "integer",
            "example": 0
          },
          "data": {
            "description": "返回数据",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubscriptionPlanDTO"
            }
          }
        }
      },
      "SingleResultSubscriptionPlanDTO": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "data": {
            "description": "返回的单条数据(object)",
            "$ref": "#/components/schemas/SubscriptionPlanDTO"
          }
        }
      },
...
      "SubscriptionPlanAddCommand": {
        "type": "object",
        "properties": {
          "planCode": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "planName": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "description": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "storageQuotaGb": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer",
            "example": 0
          },
          "maxFiles": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer",
            "example": 0
          },
          "maxFileSize": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": 0
          },
          "bandwidthQuota": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": 0
          },
          "price": {
            "format": "double",
            "description": "No comments found.",
            "type": "number",
            "example": 0.0
          },
          "isActive": {
            "format": "int32",
            "description": "No comments found.\nValidation[Min(value=0, message=isActive只能是0或1); Max(value=1, message=isActive只能是0或1)]",
            "type": "integer",
            "example": 0
          },
          "isDefault": {
            "format": "int32",
            "description": "No comments found.\nValidation[Min(value=0, message=isDefault只能是0或1); Max(value=1, message=isDefault只能是0或1)]",
            "type": "integer",
            "example": 0
          },
          "sortOrder": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer",
            "example": 0
          }
        },
        "required": [
          "planCode",
          "planName",
          "storageQuotaGb",
          "maxFiles",
          "maxFileSize",
          "bandwidthQuota",
          "price",
          "isActive",
          "isDefault",
          "sortOrder"
        ]
      },
...
      "SingleResultVoid": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "data": {
            "description": "返回的单条数据(object)",
            "type": "object"
          }
        }
      },
...
      "SubscriptionPlanDTO": {
        "type": "object",
        "properties": {
          "id": {
            "format": "int64",
            "description": "套餐ID",
            "type": "number",
            "example": 0
          },
          "planCode": {
            "format": "string",
            "description": "套餐编码",
            "type": "string"
          },
          "planName": {
            "format": "string",
            "description": "套餐名称",
            "type": "string"
          },
          "description": {
            "format": "string",
            "description": "套餐描述",
            "type": "string"
          },
          "storageQuotaGb": {
            "format": "int32",
            "description": "存储配额（GB）",
            "type": "integer",
            "example": 0
          },
          "maxFiles": {
            "format": "int32",
            "description": "最大文件数量",
            "type": "integer",
            "example": 0
          },
          "maxFileSize": {
            "format": "int64",
            "description": "单文件最大大小（字节）",
            "type": "number",
            "example": 0
          },
          "bandwidthQuota": {
            "format": "int64",
            "description": "带宽配额",
            "type": "number",
            "example": 0
          },
          "price": {
            "format": "double",
            "description": "价格",
            "type": "number",
            "example": 0.0
          },
          "isActive": {
            "format": "int32",
            "description": "是否启用",
            "type": "integer",
            "example": 0
          },
          "isDefault": {
            "format": "int32",
            "description": "是否为默认套餐",
            "type": "integer",
            "example": 0
          },
          "sortOrder": {
            "format": "int32",
            "description": "排序值",
            "type": "integer",
            "example": 0
          },
          "createdAt": {
            "format": "string",
            "description": "创建时间",
            "type": "string",
            "example": "yyyy-MM-dd HH:mm:ss"
          },
          "updatedAt": {
            "format": "string",
            "description": "更新时间",
            "type": "string",
            "example": "yyyy-MM-dd HH:mm:ss"
          }
        }
      },
...
    }
  }
}